# Make Foma export faster by mix of avoiding chain map and implement inner loop in C