… fst_lookup/_parse_foma_state.c