Some sketchy timing information (loading the Cree analyzer):


|  Original   |   C-loop    | NoChainMap  |    Both     |
|------------:|------------:|------------:|------------:|
|       1.47s |       1.25s |       1.18s |       0.84s |
|       1.55s |       1.25s |       1.18s |       0.85s |
|       1.58s |       1.25s |       1.18s |       0.85s |
|       1.62s |       1.26s |       1.18s |       0.85s |
|       1.66s |       1.26s |       1.19s |       0.85s |
|       1.66s |       1.27s |       1.20s |       0.85s |
|       1.68s |       1.27s |       1.20s |       0.86s |
|       1.77s |       1.28s |       1.21s |       0.87s |
|       1.80s |       1.29s |       1.24s |       0.88s |
|       1.84s |       1.39s |       1.28s |       0.89s |
| 1.62s–1.70s | 1.26s–1.29s | 1.19s–1.21s | 0.85s–0.87s |